fix(zcncore): added gomobile support on TransactionSchema
